remove all the warning during compilation ( thanks 2useven10 ) updated module malloc.c to the current stable version

+++ colorer/src/pcolorer2/FarEditor.cpp
@@ -698,7 +698,7 @@ const int FILTER_SIZE = 40;
         wcscpy(autofilter, prefix);
       }
     }
-    autoIncremented = wcslen(autofilter) - wcslen(filter);
+    autoIncremented = (int)(wcslen(autofilter) - wcslen(filter));
 
 
     wchar_t top[128];
